refactor: split irc and http connectors to appropriate files
