+++ internal/operator/peer/irc_connector.go
@@ -1,0 +1,11 @@
+package peer
+
+type ircPeerConnector struct {
+	serverPath string
+	channel    string
+}
+
+func (c ircPeerConnector) connectPeers(peer *Peer) error {
+	//TODO implement me
+	panic("implement me")
+}
